requirements.txt
numpy==1.24.3
pytube3
audioread==3.0.0
pip==23.1.2
ipython==8.14.0
librosa==0.10.0.post2
matplotlib==3.7.1
pydub==0.25.1
python_Levenshtein==0.21.1
PyWavelets==1.4.1
scipy==1.10.1
streamlit==1.23.1
wheel==0.40.0
vamp


